[URP] Workaround for MSAA issues on OSX when NRP RenderGraph is enabled
NOTE: THIS IS NOT AN USER FACING FEATURE, IT IS A HIDDEN RENDER PATH NOT EXPOSED TO USERS YET, THIS PR IS PART OF A SERIES OF BUG FIXES THAT WILL ALLOW US TO ENABLE THE FEATURE SO NO PUBLIC JIRAS/RELEASE NOTES ARE REQUIRED

RenderGraph Epic:
https://jira.unity3d.com/browse/URP-1297

POI:
https://jira.unity3d.com/browse/POI-817

This is a workaround for an OSX issue (reported here: https://jira.unity3d.com/browse/UUM-42825)

When changing the number of MSAA samples of the backbuffer, on OSX, the Screen API returns the wrong number of samples on the frame when the change is requested. This caused a RenderPass setup error to be printed to console on the first frame. 

This PR makes sure that, on that platform, we use the correct number of samples so no errors are displayed.

Error being fixed:
<img width="507" alt="image" src="https://media.github.cds.internal.unity3d.com/user/2433/files/5c9bb94d-f6d9-4899-8ff8-d51926f9e1b9">