[URP][RG] Framebuffer fetch
NOTE: THIS IS NOT AN USER FACING FEATURE, IT IS A HIDDEN RENDER PATH NOT EXPOSED TO USERS YET, THIS PR IS PART OF A SERIES OF BUG FIXES THAT WILL ALLOW US TO ENABLE THE FEATURE SO NO PUBLIC JIRAS/RELEASE NOTES ARE REQUIRED

RenderGraph Epic:
https://jira.unity3d.com/browse/URP-1297

POI:
https://jira.unity3d.com/browse/POI-817

https://jira.unity3d.com/browse/URP-1465  - JIRA for the task

This PR enables framebuffer fetch for URP RenderGraph implementation. It is utilised in Deferred and Decals passes. This required quite some changes, especially moving SetGlobalTexture calls to not break passes.

As well as adding new XR-specific fb fetch macros that would use Texture2DArrays for the fallback path in a similar manner that we have TEXTURE2D_X_ macros to acommodate that

The code mostly defaults to FB Fetch path except for DX12 which doesn't yet support NativeRenderPass Compiler for RenderGraph. Otherwise this is a pretty straightforward PR