BRG automatic drawcall split when using sorting positions
When sorting positions were given for a BRG draw command, the visibleCount had to be set to 1. This was because only the draw commands themselves were sorted. This PR changes this and splits the draw calls when visibleCount is larger than 1. This is a QoL improvement and a performance improvement in the same time since the same thing can be achieved without having the create individual drawCommands JIRA ID: https://jira.unity3d.com/browse/PLATGRAPH-1062