2022.3: Failure to cull VFX commands when no renderer in the CullResults
This PR fixes an issue where the VFX camera commands where not culled if no renderer was visible by a camera. 
That was because we created and registered new VFXCullResults in the callback `RendererCullingOutputReady `which was called only if there is at least one renderer in the camera's cull results.
Here, we create and register the VFXCullResults in `AfterCullingOutputReady`, which is always called when culling is over.

Parent: https://jira.unity3d.com/browse/UUM-41136
Backport: https://jira.unity3d.com/browse/UUM-41139