[Water] Fix exluders ignoring underwater
Fix exluders when camera is underwater.
Also did a bit of refactoring.

This PR is based on another one, only last commit is relevant

This exposed an issue with fog on transparents, which had no effect for regular fog but is visible on water. Previously fog wouldn't modify the alpha value of a transparent, so a transparent object with really thick fog in front would still be transparent.
Since atmospheric fog is applied behind this wouldn't cause a visible issue, but for water, this means underwater wasn't visible on transparent objects with a water exluder located behind