Fix wizard asking to migrate when not needed
Fix issue reported here https://forum.unity.com/threads/water-system-for-the-high-definition-render-pipeline.1203751/page-12#post-9153515

When downgrading from eg. 23.2 to 23.1, the hdrp asset might be at a version number that do not exist. This is not an issue technically, but currently the wizard reports that as a problem and HDRP refuses to build.

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/9ccfd4bb-17ce-4c8d-9fef-f0f3c1487abd)

It's possible to fix it by opening the asset in a debug inspector and putting an existing value, but users usually don't know about it and it's useless to force them to
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/5d6e2740-2637-4989-a15b-29933f1cbe9a)