Optimize usage of DepthFirstCollectNodesFromNode
Many users of DepthFirstCollectNodesFromNode don't care about the order of the results. They just query if an item exists in the Set. therefore, it's quicker to use a HashSet than an array, because Contains checks are faster in a HashSet.

To achieve this, I have relaxed the parameter type from List to ICollection, so either a List or a HashSet can be used, depending on what is required.