[backport] Fixed GPU crash on platforms using HLSLcc.
https://jira.unity3d.com/browse/UUM-41156
[Backport] Fixed a GPU crash on Nintendo Switch when sampling probe volume in the pixel shader.
The issue was that a structured buffer was sampled out-of-bounds (_APVResCellIndices), which is normally avoided using branching. However, when compiling for Nintendo Switch the condition is flattened, most likely using a step() pattern (common "bug" with HLSLcc). To avoid this, we need to extract the condition and force a dynamic branch.