[2023.3] Speed up HDRP Shader scriptable stripping
This PR addresses some HDRP shader stripping overhead I observed while profiling builds on a customer project.  In the HDRP stripping code, we are accessing the render pipeline settings and its resources _many_ times separately. It looks innocent: Isn't this just a property access? But in C#, a property can essentially incur whatever cost it wants, and in this case it calls into native code and does string marshalling. This is surprisingly costly when you do it over and over again. "Can't the C# compiler just optimize this for me?" Well, no. The compiler doesn't know whether your property has any side-effects, and in this case it's calling out into code that is entirely opaque to C#. Even in simpler cases, there is absolutely no expectation that this sort of cost gets optimized out.

The cost is quite pronounced:
![hdrp-shader-stripping-before](https://media.github.cds.internal.unity3d.com/user/1745/files/1f6110d9-8362-4315-b1b2-01234cb735db)

The fix is reasonably simple: We just query this information once. We're still going to look it up once per shader, which is bad, but at least it's not getting looked up thirty times or so.