[HDRP] [Water] Option to reduce scripting interactions cpu cost
Optimize water surface scripting interaction by offering an option to copy the simulation from gpu instead of doing the FFT simulation on cpu
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/15f5d96e-6f2e-4225-adda-b7ccbe9a0b66)

When this option is disabled, the simulation is always at full res because we get the data from gpu. So it gives faster and more accurate result than with the option enabled, but it will be available with a few frames delay. In practice it's not really noticeable and both options gives similarly good results, only difference is the cpu cost.

The cpu cost is now pretty much negligible