[Backport][2023.2][RPW] Added 2 callbacks for Pipeline Creation and Disposal
https://jira.unity3d.com/browse/XPIPELINE-623
https://jira.unity3d.com/browse/UUM-20646
The behavior described in the bug is actually correct. Unfortunately there is no clear way around to understand when pipeline constructed. We introduced these 2 callbacks so any subsystem interested in a point right after creation or disposal of pipeline can subscribe to it and do related logic.

We also deprecate `beginFrameRendering` and `endFrameRendering` as these 2 callbacks duplicate `beginContextRendering` and `endContextRendering`. The only diffirence that deprecated ones return Array and new ones return List. As we switch to Render method with List of Camera I think it makes more sense to keep the latter.
I also re-arranged so they behave kind of like a scope.

I set `s_CurrentPipelineType` to `k_BuiltinPipelineName` by default. Otherwise it will trigger `activeRenderPipelineTypeChanged` after re-compilation.