[URP] [CI] Fixed Linux & Switch job configs excluding HDR Output
This PR cleans up the job config and URP PostPro filter for the HDR output tests. Linux platforms and Switch do not support HDR output so I excluded the test inside the job config rather than the URP PostPro test filter in the project. This is because I needed to excluded tests from the Linux playmode too. But I messed up the flags and had to put a filter in the project too for Linux Standalone. 

Also now that UUM-43364 is fixed, we remove the special case that excluded Linux by name. The HDRColorGamutTests still won't execute on Linux but now it checks the SystemInfo properly.