[2023.2] [URP] Improve HDR Output test coverage
This PR modifies the current HDR Output tests in the URP Post Pro project so they fail if the test isn't run with HDR output (or its faking mode) enabled. This includes various fixes to make the test, now called 373_HDROutput_ColorGamut, pass on all required platforms.
On top of the 373 test, this PR adds the 374_HDROutput_PassOrder scene. This renders 4 cameras with overlay UI, each with a different pass execution order (FinalBlit, UberPost, UberPost+FinalPost, UberPost+FinalBlit) to ensure there's no breakage in executing HDR output-specific code path.

Note the reference images look washed out/a bit too bright. That's intended since the swapchain is still SDR when the HDR faking mode is active.
![373_HDROutput_ColorGamut](https://media.github.cds.internal.unity3d.com/user/2846/files/e3db9a73-b391-445f-a5e3-de0cab5194fa)
![374_HDROutput_PassOrder expected](https://media.github.cds.internal.unity3d.com/user/2846/files/5adc3bbd-8e8f-44c0-b494-4d0eba51ba91)

Finally, the RenderGraph path was fixed to pass the overlay UI texture to the final pass. Otherwise it looked too bright in playmode only with the render graph, because it was binding a default grey texture.