[SRP] Fixing Bloom Mip Bias in Screen Space Lens Flare overwriting the original Bloom mip
While making this [simple update to the HDRP Template](https://github.com/Unity-Technologies/Graphics/pull/7934), we discovered an issue on the bloom when the Bloom Mip Bias was > 0 on the Screen Space Lens Flare override. (cf [video](https://unity.gyazo.com/9b281901f2c92793cd1f06c494db6d18)). In the video, the bloom on the overexposed part shouldn't change when adding SSLF, only the flares should be added. 

Recently, a change was made to save costs. This change was basically to remove the SSLF buffer and write (adding) its result directly to the bloom buffer so that everything is composited once on the Uber shader. 

Before this PR, we were: 
- Passing the bloom mip X to the lens flare function
- Doing the lens flare stuff based on mip X
- Adding the result on the bloom mip X texture
- Writing the result back on the original bloom texture. 
Which was bad if we used a mip other than 0, since it was basically writing higher mips into the orignal "crisp" bloom texture. 

Now, after this PR, we are: 
- Passing the bloom mip X to the lens flare function **alongside the original bloom texture**
- Doing the lens flare stuff based on mip X
- Adding the result on the **original bloom texture**
- Writing the result back on the original bloom texture. 

Before => [video](https://unity.gyazo.com/f472eb289829bb1787c6c39a0636d873) (Observe the bloom on the post light)
After => [video ](https://unity.gyazo.com/d064999e4dbfb1be2bd44e9e8b4a9803)(Now the bloom on the post light doesn't change when adding / removing SSLF

This PR was also the opportunity to do 2 small refactor for better future maintenance, basically moving things from SRPs side to Core. This does not save costs, but avoid modifying things on one end and forgetting to do it on the other side: 
- First one was to move ShaderIDs related to the SSLF shader to Core and remove them from HDStringConstants.cs and PostProcessPass.cs
- Second one was to move the small computation we do with the lens flare parameters to core as well since they are the same between SRPs