[Backport][2022.3][VFX] Fix undefined behavior of SpawnerSetAttribute when an expression is connected to activation slot
With the addition of activation slots in operators, the SpawnerSetAttribute operator can sometimes have 2 parameters instead of 1, if there is an expression connected to the activation slot.
When this happens, the code assumes that the first parameter is the expression for the activation slot, and the second one is the attribute. However, this was never enforced when compiling the graph, leaving the order (and therefore the behavior) undefined, depending on the order in which they are stored by a Dictionary.
We fixed it by explicitly sorting the values in the spawner system tasks. The sorting is done by expression index, to be consistent with the sorting in the particle system, with the exception the the activation slot, that is forced to be always first.

The fact that the JIRA reproduces the issue activating "Runtime mode" is unrelated, it leads to a different order of the expressions that happens to trigger this.