[URP] Soft shadows quality level keywords
So due to the fact that soft shadow quality levels are controlled by dynamic branches, this has a huge effect on lower end platforms. This PR tries to fix that by disabling per-light quality levels and use only the global setting in the URP Asset. This affects HL/Quest, but can be expanded easily to other platforms as well.

This PR adds some tests for this as well in the shader stripping tests.

The keyword number shouldn't change too much for all platforms, as the shader stripper should remove all the new additions on other platforms and remove the old variant on the Quest/HL

Adds warning box as well when there is at least one XR provider 
![image](https://media.github.cds.internal.unity3d.com/user/735/files/45dcbb07-8665-438c-a9a2-028cc38e0621)