[Backport 2022.3][Jira # UUM-30744] Fixing miscelaneous GC Alloc issues in HDRP
Miscelanous GC fixes in HDRP. These fixes include:

Removing linq and complexity on light units validation. Light units validation was using Linq, which is full of memory allocations and very expensive in the CPU.
Instead, opting to use a simple bitmask to check wether light unit is valid or not for a certain light type. Caching also managed arrays to avoid in frame allocations.

Removing calls from GetTempTexture in ReflectionProbeTextureCache, and instead creating a custom tmp resource ticker. Calling every frame on this causes GC Alloc to be executed. It might be better to keep the texture around and instead deallocate the tmp textures after unused in several frames.

Tapping out the number of spam errors produced in ReflectionProbeTextureCache. These errors run every frame. This can cause false positive on allocations and slow down the frame rate. New code will check if the number of errors exceed a threshold, we stop showing them.