[2023.2][URP] Making sure additional lights index arrays are properly filled (UUM-46577)
Backport of #34779
Fixes UUM-46577.
The problem was that the `m_VisibleLightIndexToAdditionalLightIndex` array in `AdditionalLightsShadowCasterPass` wasn't filling out the array for all lights. It stopped if a light didn't fit in to the shadow slices (didn't cast shadows). This reproduced when a scene had a mix of shadow casting lights and non shadow casting lights.

This PR changes the loop that fills out the array so it doesn't skip lights that might not go in to the shadow slices. It also adds a new test scene to prevent future regressions.