[2023.2] Backport APV & Volumetrics Memory Access Fixes (Trunk PR #33671)
Fix OOB Memory Access in APV Logic
A previous change addressed one of the OOB memory access issues within
the APV sampling logic, but there was still another case near the bottom
of the function that could result in an OOB access. This commit adds
similar logic to that fetch operation to avoid additional issues.

Rename maxLocalIdx to maxLocalIdxPlusOne in APV
This change renames the max local index variable to better reflect what
is written into it on the CPU side. This intent of this rename is to
make it more clear why a less than comparison is the correct choice in
the APV shader code.

Fix OOB Memory Read in Volumetric Lighting Shader
This change updates the light prefiltering logic to ensure that the
object index is fully validated before it is used to read a light data
structure from memory. In some cases such as the area between the first
and second rooms in the HDRP sample scene, the first entry in the list
is not a local light which means the index is invalid for light data.