[HDRP] Added Volumetrics settings on AreaLight for path tracing.
POI: https://jira.unity3d.com/browse/POI-816 - SRP production ready

Contrary to directional and point lights, area lights do not interact with the volumetric fog with the HDRP deferred or forward renderer but do when using the path tracer. Unfortunately, the user cannot disable area lights from doing so although the option is available for other types of lights through a "Volumetrics" section in their inspector window. This was identified in [this Jira ticket](https://jira.unity3d.com/browse/FSP-1505) by the Solutions Films team.

This PR exposes this Volumetric section to area lights as well and adds a warning to inform the user that area lights volumetrics are only handled with path tracing. The settings are disabled if raytracing isn't enabled in the current HDRP asset. The volumetrics settings were already fully functional so this PR only adds some extra UI.

This will only work on rectangular shapes, as this is the only area light shape supported in the path tracer right now.

![Capture d’écran 2023-08-21 à 14 49 45](https://media.github.cds.internal.unity3d.com/user/4198/files/d683c1d6-d4a8-4a19-b63c-470281e8669d)