Make sure to bind buffers of expected size to make Metal validation h…
This PR addresses https://jira.unity3d.com/browse/UUM-45421.

The Metal validation layer checks whether buffer bindings are 'meaningful'. For example, if you bind a buffer of size 4 to a slot which points to an array of uint3s (which has size 12), then Metal validation layer will trigger because this is not 'meaningful' (myBuffer[0] will read out of bounds).

In case of [UUM-41474](https://jira.unity3d.com/browse/UUM-41474), this meant that the whole editor crashed because ProbeReferenceVolume bound a buffer which (from Metal's point of view) was not meaningful. The fix was therefore to make sure to provide a 'meaningful' uint3 buffer in the case when needToBindNeutral = true.