[HDRP] Fix fog on spherical planets
Moves the planet settings from PBR sky to visual environement.
We can now use the planet position and radius when computing the height fog to better support planets.
If planet is set to flat, behaviour is the same as before this PR.

Also refactored a bit the code to share sun light color attenuation across all effects

A next PR will add support of planet for clouds.

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/868eac13-ea8f-4116-a9a8-2e2326394f70)