[Backport 2022.3] Fix Material and Preview Prefabs light switching
Fixes an issue where material previews were rendered dark and the prefab preview would show dark until the object was moved. This issue was a result of the HDAdditionalLightData not being properly initialized when the previews were rendered.

We realized this is a regression stemming from https://github.cds.internal.unity3d.com/unity/unity/pull/20119which was solving a different problem so this fix is basically an addition to it that fixes the resulting issue of dark previews.