Fix camera position when using the light placement tool
This PR fixes the camera position when using the light placement tool.
Previously, by default, the camera was 2 meters behind the light. This was confusing and problematic when manipulating the light because a mesh could easily end-up between the light and the camera and hiding the view (E.g it was hard to place lights in interior rooms)

It also fixes another unwanted behavior. Previously when entering and exiting the tool, the light was slightly offseted, which was really problematic as this is a light placement tool. With this fix, the light stays at the right position when entering and exiting the tool. The camera is restored at it's original position, as expected.