[VFX] Independent SG Output
This PR is addressing a really common workflow issue while using ShaderGraph: if you are assigning a shaderGraph to a Particle Output Lit or Unlit, it will be the same outcome. However, if you used the Lit output, you can lost it while switching to another SRP.

This is a workflow issue coming from the wrong initial design, in this PR, we are moving all ShaderGraph Output to a dedicated independent output.

This is a good opportunity to improve the implementation and UX but also and separate old shadergraph integration (not listed since 2023.1) to the new one.

Last but not least, this PR is removing the experimental flag for ShaderGraph Output, it's now accessible to everyone 🥳 

A quick glance at the full workflow:

https://media.github.cds.internal.unity3d.com/user/42/files/5b1a7919-7e8c-4fdc-9e3c-a7918d83d53f

_This approach is the first step towards a general refactor of our output particle implementation, it focuses on ShaderGraph first but we would like to always use this composition approach in the future._