[POI-816][HDRP]Unify Build Code
This PR aims to clean all the `IPreprocessBuildWIthReport `that we have on HDRP. This allows that we are centralizing the data used for building. We gather all the data for building HDRP, and we validate that everything is correct.

The `HDRPRuntimeResouces `were having a fetch of the `ComputeShaders`, this code is duplicated in many places. I've done a utility method for that as well.

Make sure that `HDRenderPipelineGlobalSettings.Ensure()` is called from one single place when building and not all over the pre/post processors.