Stop popping the Wizard windows when the Global Settings are invalid.
JIRA ticket: https://jira.unity3d.com/browse/UUM-45694

This one is really tricky and weird. C.f the JIRA ticket for more information.

It looks like in Unity `23.1`, there was many changes in the RenderPipeline and SRP core code. The assumption that the `Standalone Profiler` is using the Built-in pipeline is not always true anymore, since it's using HDRP in this version.

Basically, the added checks verifies if the resources were successfully created. It's never true in weird context, like the Standalone Profiler, since it's a total new process. The code avoids the Wizard window to pop in those weird context. It also makes a lot of sense, since without the global settings, the Wizard content is empty and partly useless, like on this screenshot: 

![image](https://media.github.cds.internal.unity3d.com/user/1911/files/002bae23-e560-4898-959c-5569a31470ed)

This PR could also be added in the latest Trunk version if you think this safety check definitely makes sense.