[Jira # UUM-37206] Fix temporal upscaling flag to skip temporal upscalers and TAA
Adding new checkbox to allow materials to be excluded from temporal anti aliasing.
For now, this checkbox is only exposed to transparent materials since they are the only ones that can fit a stencil bit for such operation.

The materials that go through this path will generate a special mask (stencil or a bit mask in DLSS) that just performs spatial upscaling.

This checkbox is ideal for surfaces that contain texture scrolling that has to be neat and does not have velocity information.

Comparison (left is flag off, right is flag on, notice the scrolling is more clear and less blury than the one in the right):
![bluryTextureScroll](https://media.github.cds.internal.unity3d.com/user/3327/files/e1f0a283-e471-4f91-a015-f874da0915db)

![image](https://media.github.cds.internal.unity3d.com/user/3327/files/76046247-c229-4e9d-abda-dba342fdfc23)