[HDRP] Ambient occlusion mask channel effect in path tracing modifications
With regards to the path tracer, the ambient occlusion channel of a material's mask map is currently used to attenuate lighting when sampling a ray through the material BRDF, while this AO value is being ignored when sampling a ray towards a light source. This was done to limit the strength of the effect as it would apply to direct lighting. Unfortunately, this created a difference in lighting between a rectangle area light and a quad with an emissive material of similar size and intensity, as the latter, being a mesh, is hit through the BRDF sampling. As a consequence the quad's lighting is attenuated by the AO, while the area light isn't. This was initially detected by Solutions Films as described in [this Jira ticket](https://jira.unity3d.com/browse/FSP-1502).

This PR fixes the discrepancy by applying the AO effect on both light and BRDF sampling, but only for ray depth >= 2 which is equivalent to a first bounce, thereby attenuating "indirect lighting".

Part of this POI: https://jira.unity3d.com/browse/POI-816