[RPW]Fix Smoke tests CI.
Fix Smoke test job. 

With error:

`NullReferenceException: Object reference not set to an instance of an object, File: C:/build/output/unity/unity/Packages/com.unity.render-pipelines.high-definition/Editor/BuildProcessors/HDRPPreprocessBuild.cs, Line: 72`

Due to a condition check when the system was interpreting that the build was targeting HDRP but it was not.