Batch of Fixes for HQ Lines (Post-SIGGRAPH Reconcile)
POI: https://jira.unity3d.com/browse/POI-816 

### World-space Width

This PR corrects a somewhat serious impairment on an artist's ability to express the width of a line with the High Quality Line Rendering feature in HDRP. 

Up until now, the width of lines with this feature was nearly unit-less, with an artist expressing the screen-space coverage of a line on the Fragment Stage input block of Shader Graph (left below image). Now, a user is able to express the width in centimeter units on a per-vertex basis (right below image). The computation of screen coverage from this input centimeter width is now automatic, liberating the artist from the painful process of balancing screen space pixel width with alpha to get a desirable result with the High Quality Line Renderer. This unlocks instant, expected high quality results from hair DCC tools like Wētā Wig. 

This work has the greatly added benefit of finally making the image result produced by the High Quality Line Renderer completely resolution-agnostic.

![image](https://media.github.cds.internal.unity3d.com/user/1839/files/4f1265ee-b67e-4d0e-8b5b-93032971790d)

The below image is a production content example from a recent SIGGRAPH demo. The eyebrow asset came from Wētā Wig, which was encoded with a width attribute, crucial for film-grade image quality. Before, we were not able to correctly take advantage of this width attribute, now the process of using it in the high quality line renderer is as simple as connect the data to the per-vertex input slot as shown in the above image.

![image](https://media.github.cds.internal.unity3d.com/user/1839/files/7dae8b64-766b-46a3-ae1d-b5407fd6256b)

We also used this feature on another SIGGRAPH demo for the peach fuzz on this alien character. However this was just a uniform width, about 0.01cm.

![image](https://media.github.cds.internal.unity3d.com/user/1839/files/c7271d59-df10-4c13-9aa8-6310c08432b3)

### Screen Coverage LOD Mode

An additional improvement that was made to the High Quality Line Rendering feature was adding a new rendering LOD mode based on screen coverage of a hair asset.

The render LOD mechanism of HQ Lines essentially just culls strands, based on some kind of heuristic, before those strands advance too far into the software rasterizer of the HQ line renderer and unnecessarily hurt performance. Very often an artist can get a way with rendering only a certain % of strands depending on camera distance or screen coverage of a hair asset. 

The screen coverage LOD mode allows an artist to define an animation curve, where the x-axis is the ratio between the screen space bounding-box of the hair instance to the viewport. The y-axis is the percentage of strands to render based on this ratio. 

![GPUBounds](https://media.github.cds.internal.unity3d.com/user/1839/files/2d7c04f3-0318-466d-8d36-0f2f6f5b1eb0)

### Misc. Fixes and Improvements

- Replaced the persistent wave model with indirect dispatch for Fine Raster Stage.
- Improved device work saturation for Fine Raster Stage.
- Changed the binning method for detecting line/bin intersection to use the Cohen-Sutherland algorithm (more accurate)
- Introduce Native 16-bit for the Fine Raster Stage to improve occupancy. 
- Optimize a color-from/to-absorption routine.
- Various small bug fixes mentioned in the release notes.