[POI-816][RPW]Dumping in the temp folder the stripping of IRenderPipelineGraphicsSettings
In order to help `IRenderPipelineGraphicsSettings `or `IRenderPipelineResources `Devs and QA, now we are exporting into `Temp/graphics-settings-stripping.json` a file with the stripping information and which settings were transferred into the player.

This is an example of JSON:

```
{
    "totalSettings": 3,
    "totalSettingsOnPlayer": 1,
    "settings": [
        {
            "type": "RenderPipelineGraphicsSettings.StripperTests+A, Unity.RenderPipelines.Core.Editor.Tests, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null",
            "isAvailableInPlayerBuild": false,
            "strippersDefined": true
        },
        {
            "type": "RenderPipelineGraphicsSettings.StripperTests+B, Unity.RenderPipelines.Core.Editor.Tests, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null",
            "isAvailableInPlayerBuild": true,
            "strippersDefined": false
        },
        {
            "type": "RenderPipelineGraphicsSettings.StripperTests+C, Unity.RenderPipelines.Core.Editor.Tests, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null",
            "isAvailableInPlayerBuild": false,
            "strippersDefined": false
        }
    ]
}
```

Here we can see valuable information such us how many settings were transferred into the player, and how many were the original settings.

For each setting we have the `AssemblyQualifiedName`, if is one of the settings available in the player, and if there was an external stripper defined that was overriding the default behaviour of `IRenderPipelineGraphicsSettings.isAvailableInPlayerBuild`