Adding Node Reference sample pack to Shader Graph - 2022 backport
This PR adds a new pack of sample content to the Shader Graph package that users can optionally install from the Samples tab of the Shader Graph package in the Package Manager. It contains only content and assets that users can add to their project. It does not contain any changes to code or functionality in Shader Graph itself.

The Node Reference sample pack is a collection of over 140 Shader Graph assets. Instead of using these graphs for materials in their projects, users can use them as a reference to learn what each node does and how it works. Each graph represents a node that’s available in the node library, contains a description of the node, and shows how the node is used. Also included are breakdowns of how the math works under the hood and examples of what the node can do. If you want to know how to use a specific node, you can open up its reference file in these samples and see descriptions, examples, and breakdowns of that node.

Here's a prototype blog post that we plan to release along with this sample package that explains to users what it is:
https://docs.google.com/document/d/1G1TgNcAx07UO6Udgz7vt1hMOy4kgN_8CJufxqR1to_0

Our Business Intelligence Plan for the release can be found here:
https://docs.google.com/document/d/1EA2u_9J9SlI8oL63e9SBvAG9pBeNPEQmw0Az9lqwgu4

We also have a larger strategy for releasing more sample content in the coming year that is explained in this document:
https://docs.google.com/document/d/1_PiL9BKajcKxJxRU-HDchMjBB4ecPqi1lZibDspv3_s