Fix Overflow in Specular Lighting
Issue was detected within UUM-46781
With with half precision, LightingSpecular returns wrong result. PR fixes that.

Underlining issue is, when target is Android, we are using half precision. `pow(NdotH, smoothness)` function is called as `pow(half, half)`. Calculation overflows and result is shows color banding. With fix, `pow` is forced to use full precision variables to calculate result and color banding is not present.