[VFX] Custom attributes and built-in attributes in the blackboard
Jira: https://jira.unity3d.com/browse/VFXG-376

Design doc: https://docs.google.com/document/d/1VI5lVahioUiNGjyxbBhSzpMSM7z5_RXAYFPEakELUbU/edit?usp=sharing

The goal is to improve the use experience when dealing with attributes and custom attributes.

The blackboard is an existing panel where currently all properties are listed (eventually grouped in categories).

This feature will extend the role of the blackboard to also list all attributes, whether they are built-in or custom.

This will be easier for the user to discover what attributes are available.

Also, to ease the workflow of using attributes, the user will be able to simply drag&drop attributes on the graph to create a setter or a getter, depending where the attribute is dropped on the graph.

And to make it even more straightforward, the nodes dedicated to custom attributes are merged into the nodes for built-in attributes.

Of course, existing graphs will be automatically upgraded to adapt to new nodes.

On design side, the goal is to stay as close as current GTF design as possible to prepare users for upcoming migration to GTF.
![Unity_jUtvVrS7uO](https://media.github.cds.internal.unity3d.com/user/4003/files/1767fc06-1db0-469c-aeb9-f1e0772946d1)

⚠You need to build the editor because of a small change in GraphView.